… example/glue/CMakeLists.txt
… example/glue/EvacuatorDelegate.cpp
… example/glue/EvacuatorDelegate.hpp
… fvtest/gctest/configuration/evacuator_large_GC_config.xml
… fvtest/gctest/configuration/evacuator_small_GC_config.xml
… gc/CMakeLists.txt
… gc/base/Evacuator.cpp
… gc/base/EvacuatorController.cpp
… gc/base/EvacuatorController.hpp
… gc/base/EvacuatorParallelTask.cpp
… gc/base/EvacuatorParallelTask.hpp
… gc/base/EvacuatorWhitelist.hpp
… gc/base/EvacuatorWorklist.hpp
… gc/base/GCExtensionsBase.hpp
… gc/base/IndexableObjectScanner.hpp
… gc/base/ObjectScanner.hpp
… gc/base/ParallelDispatcher.hpp
… gc/base/standard/EnvironmentStandard.hpp
… gc/base/standard/Scavenger.cpp
… gc/base/standard/Scavenger.hpp
… gc/include/omrmodroncore.h
… gc/stats/ScavengerStats.cpp
… gc/stats/ScavengerStats.hpp
… include_core/AtomicSupport.hpp