Evacuator is the working name for a novel stack-based generational
collection algorithm. MM_Scavenger subclasses MM_EvacuatorController
which instantiates and controls MM_Evacuator instances. MM_Evacuator
instances are bound to and unbound from MM_Environment of gc threads
harnessed by MM_EvacuatorParallelTask.

Otherwise only minimal changes were required to plug the evacuator 
framework into scavenger. The intention is to minimize changes to
existing gencon reliability/performance/maintenance (eg, allowing
concurrent scavenger work to proceed with minimal regard to evacuator
development). 

Evacuator is disabled by default in this commit. To enable evacuator, 
so that the evacuator algorithm will run in place of 
`MM_Scavenger::workThreadGarbageCollect()`, set 
`MM_ExtensionsBase::evacuatorEnabled = true` before instantiating the 
generational collector.

Evacuator is capable of producing detailed tracing output for debugging
purposes and instrumentation. To enable this capability, and enable a 
number of debugging assertions, edit EvacuatorBase.hpp and replace 

```
#undef EVACUATOR_DEBUG
```

with

```
#define EVACUATOR_DEBUG
```

Debug tracing levels can then be controlled by exporting a environment
variable `EVACUATOR_DEBUG_FLAGS`, which is a bitmap of bit flags as 
defined in EvacuatorBase.hpp. 

Related to: #2361

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>