@charliegracie 2018-05-18
@genie-omr build all