@ktbriggs 2018-06-04
This was getting a bit stale. Rebased on eclipse/omr/master and pushed.

@charliegracie can we get an update on this?