@charliegracie 2018-06-04
Hi @ktbriggs , this is a very large PR which contains over 5kloc so it will take some time to complete the review.  This will require an Eclipse CQ to be approved before we can merge as well.  I will get the CQ opened tomorrow.