@charliegracie 2018-06-07
Due to the size of this contribution an Eclipse CQ needs to be approved . I have created the [CQ](https://dev.eclipse.org/ipzilla/show_bug.cgi?id=16676).