@ktbriggs 2018-06-18
Last commit (Enable >64 GC threads) addresses a bone of contention that was brought up in reviews. I pushed it here in order to enable a build on the farm to obtain an SDK that can run on available machines with >64 cores, which require glibc2.1X (my personal build vm builds for glibc2.4), 

As stated in the commit message, I think this commit is a good candidate for inclusion with the PR. Testing with <64 gc threads looks fine and none of this will be involved in runtimes that do nor explicity request evacuator in their runtime. Review should be pretty straightforward as many of the changes relate to different means of printing debug output involving bitmaps. 

But if this commit will significantly (>2 days) impede PR merge I will back this commit onto a new branch and continue there -- @charliegracie let me know.