@ktbriggs 2018-06-20
Rebased and pushed (no source changes). The previous continuous-integration/travis-ci/pr segfaulted on omrgctest almost immediately. Built omr from same branch/commit on a fyre vm and ran 100x grinder and 1000x grinder for built omrgctest without failure. I believe previous failure was unrelated to commit per se.