@ktbriggs 2018-06-20
Squashed 3 commits related to enabling >64 gc threads and removed #define macro that enables debug output, which should not be enabled for ci/production builds for this PR in any case. Reasons for previous travis/ci failure in gctest (segfault) were related to debug output so this should let the travis build run without failing. 

Previous travis/ci failures fo omrgctest were not reproducible on a Linux Ubuntu 16.04 vm on fyre after cloning and building this branch (see previous comment).