@ktbriggs 2018-06-21
I've completed changes relating to "Enable >64 GC threads" and rebased. Evacuator scaling up to 120 GC threads has been verified on a linux/x86-64 platform with 120 cores running SPECjbb2005 up to 120 warehouses. 

There is no apparent regression in evacuator performance for up to 8 CPUS. Typical results for SPECjbb2005 runoffs between scavenger and evacuator on 8-core linux/x86-64 look like this:

Scavenger:
```
collections: 2 global; 0 backouts; 3253 generational
idle time: 3.199% stall; 6.037% complete; .753% sync; 36542.617 gc-ms
throughput: 54067988736 allocated; 53643943432 evacuated; 36542.617 ms; 1433.577 kb/ms
fragmentation: 0 (0%) micro; 450032400 (.832%) macro; 424045304 (.785%) global
containment: 6.397% <64; 9.508% <128; 14.873% <256; 97.341% small (<= 256 bytes)
```
Evacuator:
```
collections: 2 global; 0 backouts; 3491 generational
idle time: 2.393% stall; 3.538% complete; .712% sync; 34783.991 gc-ms
throughput: 58201047400 allocated; 57708802976 evacuated; 34783.991 ms; 1620.177 kb/ms
fragmentation: 16547488 (.028%) micro; 468634864 (.805%) macro; 492244424 (.846%) global
containment: 13.682% <64; 18.655% <128; 25.583% <256; 97.486% small (<= 256 bytes)
```

More benchmarking and tuning is required to fully assess relative performance for >64 GC threads. Initial benchmarking with 120 threads showed evacuator more performant for SPECjvm2008/xml.validation but less so for SPECjbb2005. 

@charliegracie this PR is complete -- there will be no more source changes. Let me know if/when you would like to have these commits squashed.