@charliegracie 2018-06-25
The Eclipse CQ has been approved to be merged but it has not been completed.  There are currently discussing about doing an OMR release shortly.  The project can not have code from open CQs in the code base at the time of the release.  I am making some queries to see how long it will take to get the CQ completed.