@ktbriggs 2018-07-04
I added this to #2361 but I am reproducing these results here to highlight that work on this branch is piling up while waiting for this to be pulled (or flushed as the case may be). 

>Some results from preliminary benchmarking after allowing >64 GC threads and tuning for low-volume live sets (such as jbb2005). These results are from 5x2 (evacuator vs scavenger) SPECjbb2005 runs on an fyre ubuntu1604 vm with 8 cpus. Heap size is -Xms=-Xmx=1g. There no global collections except for the 2 that are always incurred with this benchmark during startup (these are small and precede any generational collections);

|metric|scavenger|evacuator|%difference|
|-----:|--------:|--------:|----------:|
|total SPEC score|142771|148396|3.9|
|total generational gc time (ms)|193536|168160|-15.1|
|average throughput (kb/ms)|1437|1721|19.8|

@charliegracie can you explain what is holding this decision (merge or flush) up? This has been dangling for >6 weeks here, and the "discussion" opened for internal review of the initial commit is >3months stale. If you do not wish to pursue this then please close it.