@ktbriggs 2018-07-10
@charliegracie in response to your questions above:

_Can you provide more details about the current set of improvements?_

Evacuator development is continuing on another branch (https://github.com/ktbriggs/omr/tree/evacuator-bench). The branch referenced in this pull request (https://github.com/ktbriggs/omr/tree/evacuator-bench) has been stable since June 18 but subject to rebasing from eclpise/master. The last rebase was June 26. Since June 18 work on the evacuator-bench branch includes:

- try to restore stack limit after each outside copy in flush instead of continuing to flush current frame
  -- flushing is triggered by presence of >0 stalled threads
- recycle survivor unused whitespace allocations into memory pool at end of each gc cycle
  -- this takes care of problematic large whitespace remainders in stack, copyspaces at end of each cycle
- changes to optimal white/work size calculations
  -- whitespace allocation sizes are now almost always maximal (128kb)
  -- workspace sizes are determined dynamically according to volume of work on evacuator worklist
- reduce default min/max work packet sizes to 50% of scavenger defaults
  -- this reduces stall times in evacuator
  -- use -XXgc scanCacheMin/MaxSize options to override default work packet size
- add breathFirstScanning to evacuator flags, ensure breadth-first scanning is supportable
  -- breadth-first scanning occurs when stack limit is set just above bottom stack frame
- add MM_ScavengerStats::countWorkPacketSize() and supporting fields
  -- produce work packet size histogram at end of each gc 
- include array header/padding in first array segment (set up in push())
  -- minor change to ensure that aggregate copied/scanned byte counts are current
- reduce volumes of work transferred from donors in findWork()
  -- receiver stops pulling before it receives more than is remaining for donor
- non-indexable object scanners preselect a hot field as first slot to return
  -- hot object referenced by preselected slot is then more likely to be copied close to parent

_How do the new changes compare?_

I ran 5 trials of gencon against each of 3 benchmarks selected for the proportions of small (<=256 bytes) objects in their workloads:

- transactional workload (~97% small objects)
- data compilation workload (~75% small objects)
- database processing workload (~50% small objects)

Each benchmark was tested with 3 gencon variants: 

- (master) scavenger built from omr and openj9 master branches (ie, without any of the changes included in this PR) 
- (scavenger) gencon + -Xgc hierarchicalScanOrdering built from omr and openj9 evacuator-bench branches
- (evacuator) gencon + -Xgc recursiveScanOrdering built from omr and openj9 evacuator-bench branches

The master variant is from a OpenJ9 build from the eclipse openj9 and omr master branches. It does not include any evacuator code and runs the production gencon algorithm. The scavenger and evacuator variants are both incorportated within the same SDK, This SDK was built from the evacuator-bench branches in my clones of eclipse omr and openj9, which were rebased on the corresponding master branches.

Results scraped from corresponding verbose gc logs are shown in the table below. The `gc-ms` statistic is the average gc time, `user-ms, duration-ms, system-ms` are totals, `gc-%cpu` is calculated as `user-ms/(user-ms + system-ms)`, and `gc-kb/ms` is average gc throughput. 

|algorithm|benchmark|gc-count|gc-ms|gc-%cpu|gc-kb/ms|user-ms|duration-ms|system-ms|
|--------:|--------:|---:|-------:|----:|------:|-------:|------:|--------:|
|master|transactional|17259|10.865|98.462|1482|187525|1163772|18176|
|scavenger|transactional|17969|11.268|98.517|1424|202483|1301032|19580|
|evacuator|transactional|17509|9.505|98.310|1697|166436|1089624|18728|
|scavenger|data compilation|8420|29.549|98.736|1640|248807|1572616|20124|
|evacuator|data compilation|8790|24.748|98.641|1962|217543|1426748|19648|
|scavenger|database processing|12958|2.607|89.645|566|33782|112024|12940|
|evacuator|database processing|12821|3.972|94.719|1106|50930|265484|14800|

For the SDK built from the evacuator-bench branch additional summary statistics are available for each nursery collection, including total bytes evacuated and proportions of reference arcs with 64-, 128-, and 256-byte cacheline containment. These internal metrics also provide a more detailed view of parallelism within nursery collections allowing stall (% of time spent waiting for work), end scan (% of time spent wating for other gc threads to complete heap scan), and sync (%time waiting for other gc threads to join at a synchronization point) percentages to be calculated. These are shown in the table below.

|algorithm|benchmark|evacuated-mb|%stall|%endscan|%sync|%<64|%<128|%<256|
|--------:|--------:|-----------:|-----:|-------:|----:|---:|----:|----:|
|scavenger|transactional|281620|5.387|3.209|0.766|6.370|9.521|15.031|
|evacuator|transactional|275883|3.999|2.881|0.735|13.234|18.226|25.224|
|scavenger|data compilation|398383|6.721|7.222|0.367|2.969|5.875|11.297|
|evacuator|data compilation|416910|7.718|4.597|0.317|7.619|15.626|25.488|
|scavenger|database processing|18668|18.598|15.319|9.089|7.652|8.914|9.867|
|evacuator|database processing|55010|7.242|8.108|5.163|26.417|29.616|31.473|

Cacheline containment and evacuation throughput were the main metrics that the evacuator algorithm is designed to improve. Substantial improvement in these metrics is evident in the results reported above. However, I cannot report any consistent and significant improvements in application performance. Please note with regard to the database processing results in the tables above that the evacuator runs evacuated almost 3X the amount evacuated in the scavenger runs with the evacuator-bench SDK.

I am in the process of completing the runs for master SDK with the data compilation and database processing benchmarks and will report results here when available.